% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cq_value.R
\name{Cq_value}
\alias{Cq_value}
\title{Finding quantificaiton vlaue for qPCR}
\usage{
Cq_value(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Numeric intergers
}
\description{
Finding quantificaiton vlaue for qPCR
}
\examples{
Cq_value(5)

}
