% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cq_value.R
\name{Cq_value}
\alias{Cq_value}
\title{Finding quantificaiton value for qPCR}
\usage{
Cq_value(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
A numeric value that shows how many cycles it takes for the DNA to amplify when ran in qPCR.
This can be used to quantify the total amount of DNA.
}
\description{
Finding quantificaiton value for qPCR
}
\examples{
bioinform::Cq_value(5)
?Cq_value
}
